Move Repository and model to module package
